refactor: Club adapter 계층 패키지 변경
