+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/club/JpaClubRepository.java
@@ -1,6 +1,6 @@
-package com.mobble.mobbleserver.infrastructure.persistence.club.core;
+package com.mobble.mobbleserver.infrastructure.persistence.club;
 
-import com.mobble.mobbleserver.domain.club.core.Club;
+import com.mobble.mobbleserver.domain.club.Club;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaClubRepository extends JpaRepository<Club, Long>, ClubQueryDslRepository {
